IA and DM compatibility
Trees should not crash in IA whenever TorrenIV is retrofitted into the world.

Dirty water of any sorts also gives you a nasty dose of rads, on the other hand tea sure does help with the situation and various soups.

Water tiles in RT is incompatible with IA and causes an odd and annoying visual glitch. Must convert tiles so characters don't fall through the floor.

Added configs for both